+++ src/main/java/chess/domain/Board.java
@@ -3,6 +3,9 @@ package chess.domain;
 import chess.domain.piece.Empty;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceFactory;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import java.util.HashMap;
 import java.util.Map;
