+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,7 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-
 public class Empty extends Piece {
     
     private Empty() {
