+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,7 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-
 public class Pawn extends Piece {
     
     private Pawn(final Color color) {
