+++ src/main/java/chess/domain/piece/PieceFactory.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import java.util.ArrayList;
 import java.util.List;
