+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,7 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
-
 public class Queen extends Piece {
     
     private Queen(final Color color) {
