+++ src/main/java/chess/domain/position/File.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.position;
 
 import java.util.Arrays;
